Refactor Open Project dialog into a separate module. Update event_init flag from number to boolean.
